modify tensor_util
